using SInt16 as audio output format. remove dsp operation from player code.

+++ src/core/tiny_player.hpp
@@ -31,7 +31,7 @@ public:
     
     PlayerViewPlatform *GetPlayerView();
     
-    void RenderAudioFrame(float *data, uint32_t frames, uint32_t channels);
+    void RenderAudioFrame(short *data, uint32_t frames, uint32_t channels);
 
 private:
     std::string file_url_;
